test led pwm alg

+++ software/src/led.h
@@ -2,7 +2,7 @@
 #define _LED_H_
 
 void LED_Init(void);
-void LED_SetColor(u32 led1, u32 led2);
+void LED_SetColor(u16 led1, u16 led2);
 void LED_Interrupt(void);
 
 #endif
